trying to fix

+++ src/components/ListingItem.jsx
@@ -31,7 +31,7 @@ function ListingItem({ listing, id, onDelete, onEdit }) {
     price,
     type,
   } = listing
-  console.log('bbb')
+  console.log('ccc')
   //console.log(id)
   return (
     <>
@@ -58,7 +58,10 @@ function ListingItem({ listing, id, onDelete, onEdit }) {
             onClick={() => {
               //e.preventDefault()
               //e.stopPropagation()
-              onDelete(id, name)
+              if (window.confirm('Are you sure you want to delete?')) {
+                console.log('you clicked yes')
+                onDelete(id, name)
+              }
             }}
 
             //onClick={onDelete(id, name)}
